+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added config, by default and optional at `~/.config/spotifatius/config.toml`. can be changed by passing `--config` to `monitor`.
+- Monitoring supports coloured formatting for polybar with config.
+
 ### Changed
 - When toggling the saved state, that is correctly reflected in the output.